backend/pyproject.toml
[tool.poetry]
name = "diploma-backend"
version = "0.1.0"
description = "Zero Trust security subsystem for a financial web platform — bachelor thesis"
authors = ["OleksandrYanchar"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = { version = "^0.29.0", extras = ["standard"] }
sqlalchemy = { version = "^2.0.0", extras = ["asyncio"] }
alembic = "^1.13.0"
asyncpg = "^0.29.0"
redis = { version = "^5.0.0", extras = ["hiredis"] }
pydantic = "^2.7.0"
pydantic-settings = "^2.2.0"
PyJWT = { version = ">=2.8", extras = ["cryptography"] }
passlib = { version = "^1.7.4", extras = ["argon2"] }
pyotp = "^2.9.0"
qrcode = { version = "^7.4.0", extras = ["pil"] }
Pillow = "^10.0.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
fakeredis = { version = "^2.23.0", extras = ["aioredis"] }
aiosqlite = "^0.20.0"
ruff = "^0.4.0"
pytest-cov = "^7.1.0"
greenlet = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ---------------------------------------------------------------------------
# Ruff — linting, import ordering, formatting, security rules
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
# E/F: PEP 8 + pyflakes  I: isort  N: naming  S: bandit security  B: bugbear
select = ["E", "F", "I", "N", "S", "B"]
ignore = [
    # S101 — assert is used in tests; acceptable in test files
    "S101",
    # S105 — hardcoded passwords: we use env vars but Ruff may flag config defaults
    "S105",
    # B008 — do not perform function calls in default arguments (clashes with FastAPI Depends)
    "B008",
]

[tool.ruff.lint.per-file-ignores]
# Tests are allowed to use assert and fixed secrets (test values, not real)
"tests/*" = ["S101", "S106", "S107"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--cov=app --cov-report=term-missing --cov-config=pyproject.toml"

# ---------------------------------------------------------------------------
# Coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source = ["app"]
concurrency = ["greenlet", "thread"]
omit = [
    "app/alembic/*",
    "app/__main__.py",
]

[tool.coverage.report]
show_missing = true
precision = 1
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]
